+++ x/marker/keeper/keeper_test.go
@@ -228,13 +228,15 @@ func TestAccountKeeperManageAccess(t *testing.T) {
 	_, err = app.MarkerKeeper.GetMarker(ctx, addr)
 	require.NoError(t, err)
 
-	// Grant access and check (fails on a Finalized marker without Admin grant)
-	require.Error(t, app.MarkerKeeper.AddAccess(ctx, user1, "testcoin",
-		types.NewAccessGrant(user2, []types.Access{types.Access_Mint})))
-	// Remove access fails for Finalized Marker without Admin grant
-	require.Error(t, app.MarkerKeeper.RemoveAccess(ctx, user1, "testcoin", user2))
+	// Manager can make changes to grants for finalized markers
+	require.NoError(t, app.MarkerKeeper.RemoveAccess(ctx, user1, "testcoin", user1))
+	require.NoError(t, app.MarkerKeeper.AddAccess(ctx, user1, "testcoin",
+		types.NewAccessGrant(user1, []types.Access{types.Access_Burn})))
 
-	// Admin can make changes to grants for active markers
+	// Unauthorized user can not manipulate finalized marker grants
+	require.Error(t, app.MarkerKeeper.RemoveAccess(ctx, user2, "testcoin", user1))
+
+	// Admin can make changes to grants for finalized markers
 	require.NoError(t, app.MarkerKeeper.AddAccess(ctx, admin, "testcoin",
 		types.NewAccessGrant(user2, []types.Access{types.Access_Mint, types.Access_Delete})))
 	_, err = app.MarkerKeeper.GetMarker(ctx, addr)